requirements.txt
Pillow==9.0.1
PyPDF2==1.26.0
typing==3.7.4.3
